Migrate MeteredInstrumenter to Byte Buddy
