+++ stagemonitor-core/src/main/resources/META-INF/services/org.stagemonitor.core.instrument.StagemonitorByteBuddyTransformer
@@ -1,2 +1,3 @@
 org.stagemonitor.core.metrics.annotations.ExceptionMeteredTransformer
 org.stagemonitor.core.metrics.annotations.TimedTransformer
+org.stagemonitor.core.metrics.annotations.MeteredTransformer
