+++ stagemonitor-core/src/main/resources/META-INF/services/org.stagemonitor.core.instrument.StagemonitorJavassistInstrumenter
@@ -1,2 +1,1 @@
-org.stagemonitor.core.metrics.annotations.MeteredInstrumenter
 org.stagemonitor.core.metrics.annotations.GaugeInstrumenter
