+++ stagemonitor-core/src/test/java/org/stagemonitor/core/metrics/annotations/MeteredInstrumenterTest.java
@@ -49,7 +49,6 @@ public class MeteredInstrumenterTest {
 	@Before
 	@After
 	public void clearMetricRegistry() {
-		MeteredInstrumenter.init();
 		Stagemonitor.getMetric2Registry().removeMatching(MetricFilter.ALL);
 	}
 
@@ -62,7 +61,7 @@ public class MeteredInstrumenterTest {
 	@Test
 	public void testMeteredAspectDefault() {
 		testObject.meteredDefault();
-		assertOneMeterExists(name("rate").tag("signature", "MeteredInstrumenterTest$TestObject#meteredDefault").build());
+		assertOneMeterExists(name("rate").tag("signature", "TestObject#meteredDefault").build());
 	}
 
 	@Test
@@ -80,7 +79,7 @@ public class MeteredInstrumenterTest {
 	@Test
 	public void testMeteredName() {
 		testObject.meteredName();
-		assertOneMeterExists(name("rate").tag("signature", "MeteredInstrumenterTest$TestObject#myMeteredName").build());
+		assertOneMeterExists(name("rate").tag("signature", "TestObject#myMeteredName").build());
 	}
 
 	@Test
